# specs/003-add-this-to/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: DPNR Landing Page API
  version: 1.0.0
  description: API for DPNR course registration, enrollment, and user management
  contact:
    email: Lital2923@gmail.com
  license:
    name: DPNR LTD
    url: https://www.BE-DPNR.com
servers:
  - url: https://api.be-dpnr.com/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
...
paths:
...
  /cohorts/current:
    get:
      tags: [cohorts]
      summary: Get current cohort information
      description: Returns the currently active cohort for enrollment
      responses:
        '200':
          description: Current cohort details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortResponse'
        '404':
          description: No active cohort found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cohorts/{cohortId}/capacity:
    get:
      tags: [cohorts]
      summary: Get real-time cohort capacity
      description: Returns current enrollment count and availability
      parameters:
        - name: cohortId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Capacity information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityResponse'

  # Enrollment Management
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
  schemas:
...
    CohortResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "December 2025 Cohort"
        startDate:
          type: string
          format: date
          example: "2025-12-01"
        endDate:
          type: string
          format: date
          example: "2026-04-30"
        maxCapacity:
          type: integer
          example: 20
        currentEnrollment:
          type: integer
          example: 15
        status:
          type: string
          enum: [upcoming, open_enrollment, full, in_progress, completed]
        location:
          type: string
          example: "Mazkeret Batya"
        schedule:
          type: string
          example: "Weekly evenings, 1.5-2 hours"
        pricing:
          $ref: '#/components/schemas/CohortPricing'
    CohortPricing:
      type: object
      properties:
        fullPayment:
          type: number
          example: 6400
        fiveInstallments:
          type: number
          example: 1360
        twelveInstallments:
          type: number
          example: 580
        currency:
          type: string
          example: "ILS"
    CapacityResponse:
      type: object
      properties:
        cohortId:
          type: string
          format: uuid
        maxCapacity:
          type: integer
          example: 20
        currentEnrollment:
          type: integer
          example: 15
        spotsAvailable:
          type: integer
          example: 5
        status:
          type: string
          enum: [available, limited, full]
        waitlistSize:
          type: integer
          example: 0
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Required field missing"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/v1/enrollments"